refactoring, save non-const selection in step
